#37 update file structure, update tests, update webpack config
