+++ src/game/grid-drawer.js
